scraper/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.1.0
python-dotenv>=1.0.0
supabase>=2.0.0
anthropic>=0.18.0
lxml>=4.9.0
